% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop-sim.R
\name{pop_sim}
\alias{pop_sim}
\title{Generate some simulated values}
\usage{
pop_sim(
  n = 20L,
  init_s = 100L,
  init_i = 2L,
  init_r = 0L,
  p_recover = 0.1,
  p_transmit = 0.4
)
}
\arguments{
\item{n}{population size to simulate}

\item{init_s}{size of initial population that is susceptible}

\item{init_i}{number infeceted}

\item{init_r}{initial R at time t}

\item{p_recover}{probability of recovering}

\item{p_transmit}{probability of transmission}
}
\value{
data frame with columns, S, I, R.
}
\description{
Generate some simulated values
}
\examples{

new_sims <- pop_sim()

plot(new_sims$S, type = "l", col = "blue", ylim = c(0, new_sims$S[1]))
lines(new_sims$I, col = "red")
lines(new_sims$R, col = "green")
}
